agent_runtime/requirements.txt
fastapi==0.128.0
google-genai==1.56.0
httpx>=0.28.0
pydantic==2.12.5
pydantic-settings==2.7.1
PyJWT==2.10.1
redis>=4.6.0
uvicorn[standard]==0.24.0
